Fix: Correct date display and adjust indentation
Corrected an indentation problem in the code and fixed a bug where the time would appear in the past because the timezone wasn't properly set up.

Since I’m not using time specifically and only care about date, I set it to UTC so there is no adjustment to the date. Even though I’m not in UTC, without a specific time, it will default to UTC, but the build system might not.

Signed-off-by: Matthew Krueger <contact@matthewkrueger.com>